Base class for 'textfield' widgets
Reduces repetitive code needed to fix MDL problems for textfield-based
components
