+++ ng2-components/ng2-activiti-form/src/components/widgets/tabs/tabs.widget.html
@@ -18,4 +18,3 @@
         </div>
     </div>
 </div>
--
